Refactor: Overall Section vue file name
