+++ src/routes.js
@@ -1,6 +1,6 @@
 // Set up our routes
 
-import Overall from './components/Main/Content/Overview/OverallSection.vue'
+import Overall from './components/Main/Content/Overview/AllOverall.vue'
 import Futures from './components/Main/Content/Futures/AllFutures.vue'
 
 export const routes = [
